Remove libjli darwin workaround
JDK-8235687 changed Contents/MacOS/libjli.dylib from a copy of libjli to
a symlink. This created issues for some applications such as Eclipse so
we changed it back to a symlink in GraalVM.

The workaround is now wrong on some JDK versions since libjli moved.
Also, those issues should now be resolved (JDK-8238225) so we can remove
the workaround.