Enhance AVXKind to return sizes for partial sizes
The getAVXKind function taking a length in elements only returns a
register when the length exactly matches the register length. For vector
registers you may want a register to hold 3 items but the appropriate
register size is 4. This change updates the API to return the correct
register size for use in the AMD64 code generator.