Conditional reflection and serialization config
Introducing conditions for reflection, JNI, and serialization configuration. The condition
can be only a type name. For example,
"condition": {
  "typeReachable": "org.graalvm.ReachableClass"
 },

will apply configuration only if type "org.graalvm.ReachableClass" is reachable.
The default condition for every reflection entry is "java.lang.Object" and as
such it will not be printed.

The Native Image agent does not emit conditions, but will fuse conditions
accordingly. Instead all config of one type, the fusion will happen by
the condition and the type.