Replaces GraalAccess.getOriginalTarget() with JVMCI.getRuntime().getHostJVMCIBackend().getTarget()
Fixes #3754
Fixes https://github.com/graalvm/mandrel/issues/287

`GraalAccess.getOriginalTarget().arch` code calls into the `GraalRuntime` interface,
for which exist 2 main implementations: `SubstrateGraalRuntime`and `HotSpotGraalRuntime`.
The problem is that, as a result of using that API, at least with OpenJDK 11.0.x (not labs),
`HotSpotGraalRuntime` and a load of `org.graalvm.compiler.hotspot`
classes end up being part of the points-to analysis and that causes the errors above, e.g.

```
$ grep HotSpotGraalRuntime used_classes_libnative-image-agent_20210906_125301.txt
org.graalvm.compiler.hotspot.HotSpotGraalRuntime
```

See NPE during build with OpenJDK reported on #3754.

I propose a change that doesn't require the use of `GraalRuntime`, hence there's no possibility for the hotspot implementation to make it to the analysis:

```
$ grep hotspot used_classes_libnative-image-agent_20210906_130032.txt
<none>
$
```