Fix DerivedConvertedInductionVariable zero extend
Zero extension was added to DerivedConvertedInductionVariable to
correctly handle some induction variables. The use of zero extension in
the stride node is not correct since the direction is encoded in the
sign and it must be preserved. This change forces the stride node to use
signed extension in all cases.