Use reflection to invoke HotSpotJVMCIRuntime.getCurrentJavaThread
https://github.com/oracle/graal/commit/deb80d4bc6f7115c756c619c45d48375d8109fdf
relies on method `HotSpotJVMCIRuntime.getCurrentJavaThread()` which was
introduced by https://bugs.openjdk.java.net/browse/JDK-8231828, but was
not back-ported to upstream OpenJDK 11.

This patch uses reflection to access the method, like in
https://github.com/oracle/graal/commit/ab692b25f0d66df602c997ebafbdd0532d5e2008,
to avoid compilation errors when using upstream OpenJDK 11 (which is
already known to not work with LibGraal due to JVMCI incompatibilities).

Closes: https://github.com/graalvm/mandrel/issues/310
Closes: https://github.com/graalvm/mandrel/issues/311