Thin launcher defaults to native library
Change the behavior of the thin launcher s.t. it will always try to load
the native library first and only fallback to JVM if that fails.
This addresses the development use-case where a distribution without
native images is built first, and a language library is only built after
the fact.