Fix incorrect stamp folding in IntegerAddExactNode
The stamp folding in IntegerAddExactNode generates an integer value
range that staturates at the max and min value on over/under flow even
though the node throws. This change simplifies the stamp generation and
produces an unrestricted stamp on over/under flow to prevent others
making incorrect assumptions about the behavior of the node.