compiler/svm: remove AMD64ControlFlow#ReturnOp as it has hotspot specific behavior
Instead, create an AMD64ReturnOp in SVM with SVM specific behavior.
To be more precise, ReturnOp emits vzeroupper if AVX is available. The
rationale for that is that it might return to the interpreter which is
SSE (that is already Hotspot specific). On SVM, switching between SSE
and AVX is explicit. Furthermore, this behaviour is incorrect for the
callee saved stub calling convention, where the caller might have values
in the upper half of the ymm registers alive.