Fix canonicalization of narrow comparisons
Currently, when canonicalizing a NarrowNode feeding a CompareNode to a
ConstantNode we try to fold the Constant across the narrow to save an
instruction. We correctly check the conversion does not affect the
constant, but we do not check the stamp ranges for the NarrowNode and
its input. This change corrects this and ensures we are not changing
semantics when doing this optimization.

(cherry picked from commit a817656b23b9d2db2fc31b30dbe5fbbd747019e0)