Handle phi while lowering ComputeObjectAddressNode
The current code requires a ComputeObjectAddressNode to be used only by
a fixed node so we can compute where to fix the object computation. In
some fuzzing tests, graphs were generated where the
ComputeObjectAddressNode could feed a ValuePhiNode. We can compute a
correct anchoring position for these specific floating nodes, so this
change adds this logic to enable these additional cases.