Handle stamp inverstion that is not possible
IntegerStamp has a function to invert a stamp across operations
including SignExtension and ZeroExtension. Because of the imprecision of
our stamp infrastructure it is possible through a series of signed and
unsigned stamp manipulations that you can arrive at a stamp that cannot
be inverted. This change checks for these cases and returns an
unrestricted stamp if the inversion is not possible.

(cherry picked from commit ca75557e151440f1c0334163f910bd5ea94edce3)