[GR-36936] Pass down ellipsis position in variadic signature to libffi
libffi offers to APIs for preparing a native call:

https://github.com/libffi/libffi/blob/5dcb741f1544c5e18c9dbf96aeb8b61cc556a616/src/prep_cif.c#L221-L235

The latter (`ffi_prep_cif_var`) is ought to be used for variadic
functions. In practice we only ever used `ffi_prep_cif` as it didn't
matter so far, because the underlying implementation only differs if
`FFI_TARGET_SPECIFIC_VARIADIC` is defined:

https://github.com/libffi/libffi/blob/5dcb741f1544c5e18c9dbf96aeb8b61cc556a616/src/prep_cif.c#L211-L215

And that wasn't relevant until darwin-aarch64:

https://github.com/libffi/libffi/blob/5dcb741f1544c5e18c9dbf96aeb8b61cc556a616/src/aarch64/ffitarget.h#L79

Bonus: It's also needed for linux-riscv:
https://github.com/libffi/libffi/blob/5dcb741f1544c5e18c9dbf96aeb8b61cc556a616/src/riscv/ffitarget.h#L66

OpenJDK support has landed recently, so we might care about
linux-riscv "soon" ;-)
https://github.com/openjdk/jdk/pull/6294

------------------------------------------------------------------------

The infrastructure to use `ffi_prep_cif_var` is actually there, but
since we didn't include the ellipsis in the signature it didn't propgate
down through the SignatureBuilder.