Add API to be able to preserve selected arguments in bytecode OSR
Bytecode OSR root node needs to be able to pass the parent frame as an
argument. For that it always replaced the original arguments array with
a 1-element array with the parent frame. Add an API that lets languages
customize how the arguments are created to be able to preserve arguments
they need while still passing the parent frame.