Add GraalTest method lookup with return type
The current method lookup functions in GraalTest allow you to refine the
lookup based on name and parameter types, but in the case of covariant
return types you can end up with bridge methods which have the same name
and parameter signature, but which have different return types. This
change adds an API that allows the return type to be specified so you
can properly select among method implementations in the presence of
bridge methods.