Fix file size reporting of produced image
Previously the total image size as reported by ProgressReporter
would report the size of the object file (.o) prior linking to
an executable as the total size. This doesn't match the size of
the actually produced final native image.

Fix the reporting by setting the reported size again after the
linking stage in NativeImageViaCC.write().

Closes #4706