Change optimization log API so that properties can be logged, improve dummy log, add missing canonicalizer log.
Make the condition to use the dummy log more efficient - the method filter can be
a part of this condition, rather than checking it repeatedly in the report method.
The unified logging API was changed so that properties are set first, so that they
are available in the report method. Log UnusedNodeRemoval in the CanonicalizerPhase.