[GR-41098] clear allocated memory after heapdump
HeapDumpWriterImpl is shown as global JNI reference, which will not be garbage collected (GC root).
When a heapdump is written, some fields are set, which contain a lot of memory. Those fields are
not null'ed after the .hprof file has been written. This will result in lost memory, which is quite
expensive in some cases. The fields can not be GC'ed, because they are referenced from
`HeapDumpWriterImpl` and `HeapDumpWriterImpl` is a GC root, which means it will not be GC'ed. This
fix nulls set fields after the .hprof file has been written in order to be GC'ed.

Issue-Id: GR-41098