[GR-41429] Rename rpath folder to avoid argument handling confusion in ld64
ld64 supports response files [1], that is, it will expand an argument prefixed with '@' with its file content. Consider this example:
```
$ cat hello.commands
my_arg1
my_arg2
$ ld @hello.commands
ld: file not found: my_arg1
```

This expansion is not context sensitve. For tests around rpath, we had this situation:

```
graalvm-native-clang -Wl,-install_name,@rpath/libFoo.dylib \
    [...] \
    -shared -o rpath/libFoo.dylib \
    sulong/tests/com.oracle.truffle.llvm.tests.linker.native/rpath/FooA.c
```

`-install_name @rpath/libFoo.dylib` will be passed to ld64 as arguments. According to the expansion described above, the linker will thus replace `@rpath/libFoo.dylib` with its content (that's why we saw the magic header of dylib [2] via `otool`). If it doesn't exist, the argument will be kept as-is.

Previously the latter was the case, and everything was fine. However, by enabling parallelism for GNU make projects on Darwin recently [3] plus some luck in terms of racing, it can happen that `rpath/libFoo.dylib` does indeed exist on the file system and the linker will expand the argument as specified.

[1] https://github.com/apple-opensource/ld64/blob/e28c028b20af187a16a7161d89e91868a450cadc/src/ld/ResponseFiles.cpp#L68-L73
[2] https://github.com/llvm/llvm-project/blob/56c6b612aed1b4bf691a2a5ae731c09b18652ba5/llvm/lib/BinaryFormat/Magic.cpp#L132-L134
[3] https://github.com/graalvm/mx/commit/1260fd82ffb8f84d3dbe0e0b88ee93fc345f479d