Reset JDKs cgroup subsystem instances
This is needed in order to get correct memory limits detected
for some code paths (one of which is JFR).

Reset containerMetrics and initializationTriggered static
variables in JDKEvents class so as to not carry over build-time
information to runtime.

Similarly, reset RequestEngine.entries list to an empty list
so that initialization doesn't re-add hooks added at build time.

Closes: #5396