More user-friendly reporting of internal errors.
In case of internal errors, the build process generates an error report
and no longer shows daunting stack traces. Instead, it fails with a
clear message that tells users how to proceed: inspect the error report
and, if unable to resolve the problem, file an issue with the error
report.

Inspired by HotSpot, the default filename for error reports is
`./svm_err_<timestamp>_pid<pid>.md`.