Shade JLine 3.23 again
We need to use a fork of JLine3, because of incompatibilities with
native-image. Since we are forking anyway, we can also keep the
shading.

Note that now the shaded fork also excludes all the unused JLine3
components, including Jansi terminal provider, which was broken due
to JNI linkage errors (due to relocation of Java classes, but missing
relocation of the native symbols, see GR-46271).