Fix compilation error during graph decoding when the inlining log is enabled.
The inlining log is in a default update scope during decoding. However,
some invokes may get replaced at the same time, and they try to open a
scope of their own. Only one scope at a time can be open. Fixed by not
opening the inner scope if we are already in a default scope.