[GR-43389] [darwin-amd64] Workaround for buggy ld64 versions
Some versions of `ld64` cannot deal with DIRECT8 relocations in the `.text` section. Approximately this is since version 820.1 (Xcode 14). Starting with Xcode15 beta3 the default linker has been replaced with "the new linker" (version 902.11) which does not suffer from this bug anymore.

However, Xcode also ships a `ld-classic` which reassembles "the old linker" and is still affected by this bug (and it of course prints the same version number).

See some more in-depth analysis of this ld64 bug in https://openradar.appspot.com/FB11942354

The workaround: Instead of emitting the address of a HostedMethod inlined, it will be put in the data section and thus requires a memory load to obtain it.

(cherry picked from commit e30698b53f00a021fce8e7be17da960fcb4a262a)

resolve conflicts

resolve errors

backport fixup