Treat empty stages as successful
The NativeImageVM is still invoked, even if the requested stage should
not run (e.g. `instrument-image` in a non-PGO benchmark).
In that case, the config sets it's `stage` field to None
To satisfy the requirement of a success pattern in the output, we also
print a message when a stage is skipped