BytecodeParser: Do not pop values from FrameState before resolving types/fields
If, during handling of an instruction, we come across an unresolved type/field, we generate an unconditional deoptimization. In such cases we should maintain the values that are present on the stack intact.
However, the `handleUnresolvedX` functions in BytecodeParser encouraged popping values from the stack, even though the obtained values were never used. This PR removes ValueNode parameters from `handleUnresolvedX` functions and reorders some function calls to ensure that no values are popped from a FrameState's stack until we're certain we won't emit a deopt.