When buildingDFA, use deduplication instead of insideLoops to ensure termination
This is used only for flavors which need extraEmptyLoopIterations. By
ignoring insideLoops, we can have as many extra empty loop iterations as
is needed and termination is ensured by deduplication. In order for
deduplication to work, we have to bound the number of quantifier guards
that can be present in a transition. We do this by further normalizing
the enterZeroWidth quantifier guard.

(cherry picked from commit b0dbd8852305f8f976bac3135455a556b0af1eed)