Only produce JMH performance datapoints in run stage
The JMH harness generates a JSON file with the measurements and the
bench suite, unconditionally, uses it to produce datapoints.
If we run a single stage (e.g. `image`), the bench suite will read the
`jmh_result.json` file (assuming that file was produced in a previous
stage, e.g. `agent`, otherwise it will simply crash) and use it to
produce datapoints.
Now a benchmark run of the `image` stage produced completely unrelated
benchmark data.

Because of that, we hook into get_jmh_result_file and only return a
valid path in the run stage