Ensure shaded dependencies are downloaded before the shading task
Build tasks should never download their dependencies, the build task of
the library must take care of that to avoid concurrent downloads.
Concurrent downloads are only a waste of bandwidth in most situations
but on windows they can lead to failures.
Instead, ensure that there is a build dependency from a shaded library
project to its shaded dependencies.
Also, dependencies should not be resolved in constructors but in
`resolveDeps`.