Espresso: Workaround issue with merged exits for merge-explode loop
Loop exits are merged for merge-explode loops. As a result, as soon as a
branch is going to exit the interpreter loop (i.e., a branch that ends
in a return or a throw that isn't caught in the interpreter loop), we
can't rely on the bci or top being a constant.
As a workaround, we force the interpreter to go around the loop once
more to avoid the loop exit in those branches.