Remove deleting of c.o.t.api.library.LibraryFactory.ResolvedDispatch#REGISTRY entries on cleanup
When tests are run with c.o.s.enterprise.test.runner.UnitTests (svmtest)
all test classes are loaded by NativeImageClassLoader and that
loader instance is used throughout all the subsequent image builds that
happen as part of a UnitTests run. Entries into ResolvedDispatch#REGISTRY
are put there by static initializers of generated classes. Because of
the NativeImageClassLoader reuse for all image builds this only happens
once per such class and removing the entries on TruffleBaseFeature#cleanup
would make them unavailable in a subsequent tests that also rely on the
entries in ResolvedDispatch#REGISTRY.