Espresso: when continuations are enabled, ensure lambdas are always serializable.
Note that this does not enable the use of ObjectOutputStream with lambdas - in that case the user must still opt-in via the standard Java interface intersection syntax to make javac generate the synthetic `$deserializeLambda$` method. However, in cooperation with an alternative serialization engine that doesn't depend on that synthetic method, this feature does allow transparent serialization of lambdas.