[GR-58575] SubstrateVM PLT/GOT Feature
Introduces an additional level of indirection for calls where a GOT (Global Offset Table) is an array of method pointers and PLT (Procedure Linkage Table) is a collection of small stubs.  With this feature enabled, direct calls are emitted as indirect calls through the GOT. The virtual table is filled with PLT stubs instead.

While inspired by ELF, no ELF mechanisms are used in the implementation.

Example usecase: Hijack code execution on call boundaries to diverge execution from AOT code to an interpreter.

Contributors:
- Aleksandar Gradinac: Initial implementation on linux-amd64.
- Marko Spasic: Miscellaneous improvements.
- Bernhard Urban-Forster: Support for linux-aarch64, darwin-aarch64 and darwin-amd64.
- Alfonso² Peterssen: Support for windows-amd64.

Co-authored-by: Aleksandar Gradinac <aleksandar.gradinac@oracle.com>
Co-authored-by: Marko Spasic <marko.spasic@oracle.com>
Co-authored-by: Alfonso² Peterssen <alfonso.peterssen@oracle.com>