Increase max heap size of Native Image builder to match Quarkus' CI
Since https://github.com/quarkusio/quarkus/pull/43862 the quarkus
nightly builds have started failing (stopped by the watchdog) in the
analysis phase of the micrometer-prometheus integration test. The reason
appears to be an increase in the memory usage of the Native Image
builder due to the aforementioned change.

Since Quarkus CI is already using 6g as the max heap size. It makes
sense to align this workflow with that limit.