Refactor ConditionalFlow
This commit updates ConditionalFlow to use the same pattern as PrimitiveFilterTypeFlow and BooleanPrimitiveCheckTypeFlow - use eval function, which is called upon any input signal. The only difference here compared to the two flows mentioned earlier is that ConditionalFlow has both observer and use input edges, so there are more methods to override.