Register custom javac arg processor for JDK builds without JMODs
With JEP 493, part of JDK 24, it's possible to have JDK builds without
the `jmods` folder. Currently substratevm builds assume `jmods` are
always present. This patch adds an extension to mx so as to produce
custom javac args when specific substratevm dependencies get compiled.

All of this is activated only when `--no-jlinking` option is being
used.

This patch depends on (which adds the needed abstractions to mx):
https://github.com/graalvm/mx/pull/287