Separate immutable and mutable pool constant interfaces
The goal of the refactoring is to remove all methods that can't be
implemented for resolved entries: immutable entries have all the
accessors needed to look at constant pool entries syntaxically while
resolved constant pool entries have all the accessors needed to look at
the resolved version. This remove a lot of "trap" methods that can't
always be called depending on the type of the constant pool entry and
lets that be checked at compile time.
A `isSame` method was also introduced to serve a comparison that used to
be done by comparing a string representation of the entry.

Minor:
* Remove useless override in BootstrapMethodConstant.Indexes
* Return byte from `ConstantPool.Tag.getValue()`. This avoids a cast to
`byte` at usages.
* Use modern switch in `ConstantPool.Tag.fromValue`