Fix last dacapo warmup iteration having the wrong iteration number
For N iterations, dacapo runs N-1 warmup iterations, which print
'completed warmup X', as well as a final iteration, which prints
'PASSED'.
The 'warmup' metric should capture all those values in order.
Before, it had a separate rule to capture the pattern using 'PASSED' for
the warmup, which used totalIterations (the value passed to -n) to
determine what the iteration of that measurement is supposed to be.
Using the number passed to -n to determine what number the last
iteration should be is fragile beause that captures the orgininal -n
value and not the replaced -n value for agent and instrumentation runs.