Remove sulong dynLink/cycle test.
This test is testing the loading order for DSOs with cyclic
dependencies. In glibc 2.35, the loading order changed. That means this
test doesn't work anymore on systems with glibc >= 2.35, since the
reference executable now outputs a different order than Sulong. Since
glibc doesn't considers this order stable, it doesn't make much sense
to test for it in Sulong.