[GR-65608] Improvement of Strong/WeakHandles
HandleStorage is an abstract class which lays the basis for an array-based handles implementation. It has two parameters: the type T of objects which are handlified and the type of the Reference REF of the objects which is stored.
Additionally it supports two different modes to free handles. Explicit freeing via a freeHandle method or implicit freeing where unused handles are lazily collect when needed.
WeakHandles is the implementation of HandleStorage for objects of type T and WeakReference<T> as storage REF. It is in implicit free mode.
StrongHandles stores strong reference to objects of type T and frees explicitly.