Networking Support for EspressoLibs
We added an option to disable networking on the user side. Then if we reach networking classes we always check if its enabled from the user and the Truffle side.
Generally we use the host system to implement all networking functionalities by associating a fd with the appropriate channel (eg ServerSocketChannel) in TruffleIO.
Then in the native methods we can use the provided fd to retrieve the corresponding host channel to implement the semantics.