Fixup the --future-defaults description
Improve verification of options: the list of options in help text is guaranteed to contain all options.

Mentions that it should be used with `--future-defaults=all` in the first sentence.

Removed enumeration from the list as it is not relevant and options are easier to read and copy/paste.

Option misuse now prints the whole option help text.