Fix auto-generated externs conflicting with closure externs
For every `@JS.Import`, Web Image generates an externs definition for
that type. If it is a JS built-in type like `Map`, the closure compiler
will also have an externs definition for it, which results in a multiple
declarations error.

We suppress those warnings in our generated externs and put the closure
externs first so that the generated ones comes after and the suppression
actually works