Fix: stop overriding countsTowardsStackTraceLimit.
The default `countsTowardsStackTraceLimit` impl invokes `isInternal`, which used to force materialization of sources in the Bytecode DSL.
To prevent this eager materialization, we overrode `countsTowardsStackTraceLimit` to `true` when the user did not implement it.
This is incorrect: the user could override `isInternal`, but the `countsTowardsStackTraceLimit` override would ignore it.
Since eager materialization no longer occurs, we can stop generating an override for this method.