Crema: add resolution for INVOKEDYNAMIC entries
In the `RuntimeInterpreterConstantPool`, INVOKEDYNAMIC entries are resolved
to a `ResolvedInvokeDynamicConstant`.
Those hold information about the bootstrap method and a list of call
site links. Call site links can either be successful or failed.
Successful call site links contain an invoker method as well as an
appendix.
The index of the call site link is patched into the 2 empty bytes of the
4 bytes CPI of invokedynamic bytecodes.