Revert to exnref in ValueType, forbid exnref in JS global helpers
According to the wasm JS API spec, exnref is not a valid ValueType.
However, we will need to enforce that the types v128 and exnref never
cross the wasm and JS boundary and we will need to do so in JS (we allow
v128 values crossing from wasm to the spec test harness and other
embedders). Since we use ValueType as a simple form of type reflection,
we will need to have exnref so that GraalJS can detect exnref in wasm
type signatures.