Replace InstructionImmediateEncoding#explicit with InstructionImmediate#dynamic
The way an immediate is supplied is unrelated to the encoding itself.
Moving this flag ensures methods like doEmitInstructionXYZ have a single overload, whereas before they could have multiple overloads depending on the explicit-ness of the immediates (this was fragile).

I also opted to use the terminology "dynamic" to emphasize that it is written at run time and not supplied at build time.