Fix WasmValidator being too lenient
It would propagate up the end of a block being unreachable (and thus
satisfying all requirements on return types) to its parent block.
The spec does not require that, validation assumes the inner block can
always return and the outer block is supposed to have instructions that
deal with that.

Our generated Wasm text format also did not pass validation with tools
like wasm-tools and the reference implementation (the binary did because
binaryen did the right thing).
Even though the try_table block always ended with a jump, the outer
block (the target of the catch clause) required its own (unreachable)
instruction after the try_table block to pass validation.