Allow coercion between all JSObjects and their subtypes
JSObject is just a wrapper around a JS object, coercion is
straightforward and just rewraps the JS object in a new wrapper.

This requires that all reachable JSObject subtypes are marked as unsafe
allocated because now just having access to the types Class instance
allows you to create it through coercion, without the analysis seeing it

This also simplifies some code generation, we no longer have to store
the facade mappings at runtime because all facades are valid for all
types.