Move the SIMD implementation from the api package to the vector package
The original plan was for the GraalWasm SIMD internals to be
intrinsified directly by Graal and so the method signatures were to be
part of the GraalWasm-Graal API and hence they were put in the api
package. With the use of the Vector API, this is no longer the case, so
I am cleaning up and moving the SIMD implementation to its own package.