The -H:+CompatibilityMode for Native Image
The goal of this mode is to remove native-image implementation-specific behavior: no build-time initialization on classpath, no custom system properties for native image, no substitutions on classpath, no user features, future defaults enabled by default.

It still does not modify the key restrictions of Native Image which are related to dynamic access and run-time class loading.

Also, this mode still does not disable flags such as --features and --initialize-at-build-time from the command line. This will be discussed separately and done in a separate PR.

To get the same behavior as the underlying language it is recommended to use this flag with:

  native-image -H:+CompatibilityMode -H:Preserve=all -H:+RuntimeClassLoading App

And to run it with:

  ./app -Djava.home=<path-to-java-home> -Djava.class.path=<cp> -Djdk.module.path=<module-path>